+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changes
+
+- [#4901](https://github.com/ignite/cli/pull/4901) Enable automatic switching to daemon mode when not a terminal.
+
 ## [`v29.9.0`](https://github.com/ignite/cli/releases/tag/v29.9.0)
 
 ## Features